fossilizer: refactor files
